refactor: 제출 버튼 문구 변경 및 전화번호 포매팅 함수 유틸 분리
