+++ src/util/ts/formatPhoneNumber.ts
@@ -1,0 +1,13 @@
+export default function formatPhoneNumber(phone: string) {
+  const digits = phone.replace(/\D/g, '');
+
+  if (digits.length <= 3) {
+    return digits;
+  } else if (digits.length <= 7) {
+    return `${digits.slice(0, 3)}-${digits.slice(3)}`;
+  } else if (digits.length <= 11) {
+    return `${digits.slice(0, 3)}-${digits.slice(3, 7)}-${digits.slice(7, 11)}`;
+  } else {
+    return `${digits.slice(0, 3)}-${digits.slice(3, 7)}-${digits.slice(7, 11)}`;
+  }
+}
